removed code smell

+++ spring-project/src/main/java/attsd/exam/spring/project/services/RestaurantService.java
@@ -36,7 +36,8 @@ public class RestaurantService {
 
 	public void delete(BigInteger id) {
 		Restaurant r = restaurantRepository.findById(id).orElse(null);
-		restaurantRepository.delete(r);
+		if (r != null)
+			restaurantRepository.delete(r);
 	}
 
 	public void deleteAll() {
